+++ packages/map-template/CHANGELOG.md
@@ -5,7 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.88.0] - 2025-12-09
+## [1.88.1] - 2025-12-11
+
+### Fixed
+
+- Fixed search results sorting to be numeric instead of lexicographic
+
+## [1.88.0] - 2025-12-11
+
+### Added
+
+- Do not show floor in location info when venue has only one floor.
+
+## [1.87.15] - 2025-12-09
 
 ### Added